refactoring2: arcade & document
