+++ backend/open_webui/retrieval/upstage_parser.py
@@ -5,6 +5,8 @@ from typing import Optional, List
 
 from langchain_core.documents import Document
 from open_webui.models.users import UserModel
+from open_webui.config import RAG_EMBEDDING_PREFIX_FIELD_NAME
+from open_webui.env import ENABLE_FORWARD_USER_INFO_HEADERS
 
 log = logging.getLogger(__name__)
 
@@ -220,3 +222,46 @@ async def download_and_merge_results(result_metadata):
     # merged_html += "</body></html>"
 
     # return merged_html
+
+
+
+def generate_upstage_batch_embeddings(
+    model: str,
+    texts: list[str],
+    url: str = "https://api.upstage.ai/v1",
+    key: str = "",
+    prefix: str = None,
+    user: UserModel = None,
+) -> Optional[list[list[float]]]:
+    try:
+        json_data = {"input": texts, "model": model}
+        if isinstance(RAG_EMBEDDING_PREFIX_FIELD_NAME, str) and isinstance(prefix, str):
+            json_data[RAG_EMBEDDING_PREFIX_FIELD_NAME] = prefix
+
+        r = requests.post(
+            f"{url}/embeddings",
+            headers={
+                "Content-Type": "application/json",
+                "Authorization": f"Bearer {key}",
+                **(
+                    {
+                        "X-OpenWebUI-User-Name": user.name,
+                        "X-OpenWebUI-User-Id": user.id,
+                        "X-OpenWebUI-User-Email": user.email,
+                        "X-OpenWebUI-User-Role": user.role,
+                    }
+                    if ENABLE_FORWARD_USER_INFO_HEADERS and user
+                    else {}
+                ),
+            },
+            json=json_data,
+        )
+        r.raise_for_status()
+        data = r.json()
+        if "data" in data:
+            return [elem["embedding"] for elem in data["data"]]
+        else:
+            raise "Something went wrong :/"
+    except Exception as e:
+        log.exception(f"Error generating openai batch embeddings: {e}")
+        return None
